gpt-oss-20b-red-teaming/requirements.txt
# Core dependencies for OpenAI gpt-oss-20b Red-Teaming Challenge
# Model interaction and API
openai>=1.50.0
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.30.0

# Data analysis and manipulation
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Jupyter and notebook support
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Testing and automation
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Text processing and NLP
nltk>=3.8.0
spacy>=3.7.0
textstat>=0.7.0
rouge-score>=0.1.2

# HTTP requests and API interaction
requests>=2.31.0
aiohttp>=3.8.0
httpx>=0.24.0

# JSON and data validation
jsonschema>=4.19.0
pydantic>=2.4.0

# Logging and monitoring
loguru>=0.7.0
wandb>=0.15.0

# Security and crypto
cryptography>=41.0.0

# Development tools
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Environment and configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Progress tracking and CLI
tqdm>=4.66.0
click>=8.1.0
rich>=13.5.0

# Model evaluation metrics
scikit-learn>=1.3.0
torchmetrics>=1.0.0

# Memory and performance monitoring
psutil>=5.9.0
memory-profiler>=0.61.0
